提交代码，并修复已知bug

+++ String/String.h
@@ -66,8 +66,8 @@ public:
 	void Append(const char* str);
 	String& operator+=(char ch);
 	String& operator+=(const char* str);
-	size_t Find(char ch, size_t pos = 0);
-	size_t Find(const char* str, size_t pos = 0);
+	int Find(char ch, size_t pos = 0);
+	int Find(const char* str, size_t pos = 0);
 	void Insert(size_t pos, char ch);
 	void Insert(size_t pos, const char* str);
 	void Erase(size_t pos, size_t len = npos);
